Tolerate package import without API key, refactor code structure
