+++ agent/util.go
@@ -1,0 +1,19 @@
+package agent
+
+import (
+	"fmt"
+	"os"
+	"strconv"
+)
+
+func GetBoolEnv(key string, fallback bool) bool {
+	stringValue, exists := os.LookupEnv(key)
+	if !exists {
+		return fallback
+	}
+	value, err := strconv.ParseBool(stringValue)
+	if err != nil {
+		panic(fmt.Sprintf("unable to parse %s - should be 'true' or 'false'", key))
+	}
+	return value
+}
